requirements.txt
requests>=2.31
beautifulsoup4>=4.12
playwright>=1.40
python-dotenv>=1.0
geopy>=2.4
